requirements.txt
fastapi==0.135.1
keyring==25.7.0
pwdlib==0.3.0
pydantic==2.12.5
pyjwt==2.11.0
pytest==9.0.2
pytest-asyncio==1.3.0
rich==14.3.3
tortoise-orm==1.1.6
uvicorn==0.41.0